:sparkles: get lol match, matchId, lolUserInformation

+++ app/src/main/java/com/api/study/riot_api/data/model/dto/MatchInformationDto.kt
@@ -2,12 +2,12 @@ package com.api.study.riot_api.data.model.dto
 
 data class MatchInformationDto(
     val dataVersion: String,
-    val gameCreation: Int,
+    val gameCreation: Long,
     val gameDuration: Int,
-    val gameEndTimestamp: Int,
+    val gameEndTimestamp: Long,
     val gameMode: String,
     val gameName: String,
-    val gameStartTimestamp: Int,
+    val gameStartTimestamp: Long,
     val gameType: String,
     val gameVersion: String,
     val mapId: Int,
